feat: Comprehensive FastLanes encoding strategy refactoring and performance optimization
Major Changes:
- Renamed TransposeVector to TransposeFieldEncodedAndCompressedVector for clarity
- Renamed GroupedVector to GroupedFieldEncodedAndCompressedVector for clarity
- Added TransposeFieldEncodedBlockCompressedVector with block-level compression
- Added GroupedFieldEncodedBlockCompressedVector with block-level compression
- Optimized group size from 64D to 32D for better SIMD utilization

Performance Improvements:
- Fastest decode: TransposeBlockCompressed + None at 16.64ms
- Best compression: TransposeBlockCompressed + Zstd at 1.57x ratio
- Fastest encode: GroupedFieldEncoded + None at 23.84ms
- Best balance: GroupedFieldEncoded + Zstd (1.56x, 26.63ms encode, 16.85ms decode)

Bug Fixes:
- Fixed encoding/decoding mismatch in GroupedBlockCompressed strategy
- Resolved CompressionContext API usage issues
- Fixed FastLanes encoder/decoder API calls

Documentation:
- Updated ENCODING_PERFORMANCE.adoc with latest benchmark results
- Added comprehensive technical insights and performance metrics
- Documented format markers (GV, GB, TV, TB) for strategy identification

+++ src/core/bloom/mod.rs
@@ -708,10 +708,10 @@ impl SstableBloomFilter {
         buffer.write_all(&self.stats.metadata_queries_saved.to_le_bytes())?;
 
         // Write data lengths and data
-        buffer.write_all(&(self.key_filter_data.len()).to_le_bytes())?;
+        buffer.write_all(&(self.key_filter_data.len() as u32).to_le_bytes())?;
         buffer.write_all(&self.key_filter_data)?;
 
-        buffer.write_all(&(self.metadata_filter_data.len()).to_le_bytes())?;
+        buffer.write_all(&(self.metadata_filter_data.len() as u32).to_le_bytes())?;
         buffer.write_all(&self.metadata_filter_data)?;
 
         Ok(buffer)
